Views.py to modules
