+++ libriscan/biblios/models/__init__.py
@@ -1,3 +1,4 @@
+__all__ = ["base", "documents", "organizations", "users"]
 from .users import User, UserRole
 from .organizations import Organization, CloudService, Collection, Series
 from .documents import Document, DublinCoreMetadata, Page, TextBlock
