+++ libriscan/biblios/views/__init__.py
@@ -1,0 +1,36 @@
+__all__ = ["base", "organizations", "documents", "words"]
+from .base import index, scan
+from .organizations import (
+    organization_list,
+    OrganizationDetail,
+    OrganizationUpdate,
+    CollectionCreate,
+    CollectionUpdate,
+    collection_detail,
+    SeriesCreateView,
+    SeriesDetail,
+)
+from .documents import (
+    DocumentList,
+    DocumentDetail,
+    DocumentCreateView,
+    DocumentDeleteView,
+    MetadataDetail,
+    MetadataUpdateView,
+    PageCreateView,
+    PageDetail,
+    handle_upload,
+    extract_text,
+    export_pdf,
+    export_text,
+    export_xml,
+    check_words,
+)
+from .words import (
+    update_word,
+    update_print_control,
+    update_text_type,
+    textblock_history,
+    revert_word,
+    merge_blocks,
+)
